Ensure only one scheduler runs a job at a time
In a multi-server environment, the scheduler may execute multiple times. Patch limits the effect by checking to see the update succeeded in marking any rows for update and only continuing if it has. The assumption is that if it has not marked any rows for update then there is no work to be done. I.e. another process has taken over.